spi-ingenic-nouveau: war for X1000 TX FIFO errata
The TX FIFO must not contain more than one word when working in full-duplex
DMA mode. Otherwise it dies randomly when PDMA has more than one users.

I was trying to implement RX only mode but didn't success. It's not
implemented in Ingenic's 4.4/5.10 kernel as well, so it's probably also
an errata.